refactoring
Co-collaborator: Wayne Tsai
