+++ .env
@@ -1,0 +1,1 @@
+POSTGRES_URI = "postgres://eylrwzmt:oZ-T6w_cEIg_2i9agFo6RHf3OpqA82NU@heffalump.db.elephantsql.com/eylrwzm?ssl=true"
