--- components/DragAndDrop.tsx
@@ -1,22 +1,0 @@
-//import statements
-// import React from 'react';
-import { React } from '../deps.tsx';
-
-export default function DragAndDrop(){
-// state
-
-  // const dragAndDropStyle = { 
-  //   gridArea: 'dd',
-  //   backgroundColor: 'rgb(230, 203, 27)',
-  //   border: '2px solid white',
-  //   fontSize: '30px',
-
-  // } as const;
-
-
-  return (
-    <main >
-      <p>Drag n Drop</p>
-    </main>
-  );
-}
