--- components/Preview.tsx
@@ -1,20 +1,0 @@
-//import statements
-// import React from 'react';
-import { React } from '../deps.tsx';
-
-
-export default function Preview(){
-  // state
-  // const previewStyle = { 
-  //   gridArea: 'side',
-  //   backgroundColor: 'rgb(255, 123, 0)',
-  //   border: '2px solid white',
-  //   fontSize: '30px',
-  // } as const;
-
-  return (
-    <main >
-    <p>Preview</p>
-    </main>
-  );
-}
