--- components/customization/pages/Routing.tsx
@@ -1,10 +1,0 @@
-import { React } from '../../../deps.tsx';
-
-const Routing: any = (props: any) => {
-
-  return (
-    <div>Routing page in react router</div>
-  )
-}
-
-export default Routing;
