+++ deps.tsx
@@ -13,8 +13,3 @@ export * as serve from "https://deno.land/std@0.140.0/http/server.ts";
 
 export * as ReactDOM from "https://esm.sh/react-dom@18.2.0";
 export { React, BrowserRouter, HashRouter, Routes, Route, Link, useNavigate };
-// export { serve };
-
-//Twind
-// import { setup, tw } from "https://esm.sh/twind@0.16.16";
-// import { getStyleTag, virtualSheet } from "https://esm.sh/twind@0.16.16/sheets";
