+++ server/routes/projectRoutes.ts
@@ -1,6 +1,6 @@
 import {  Router, RouterContext } from "oak";
-import db from '../db.ts'
-import projectController from '../controllers/projectController.ts'
+import db from '../utils/db.ts';
+import projectController from '../controllers/projectController.ts';
 import {authorize} from '../controllers/authorize.ts';
 
 const router = new Router();
@@ -11,4 +11,4 @@ router
   .post('/home/save', authorize, projectController.saveProject)
   .post('/home/delete', authorize, projectController.deleteProject)
 
-  export default router;
+export default router;
