+++ src/utils/types.ts
@@ -1,30 +1,17 @@
-//Frontend
-
 export interface IHtmlElement {
   id: number;
   element: string;  
   inputText: string;
-//   setInputText?: (inputText: string) => void;
   texAlign: string;
-//   setTextAlign?: (textAlign: string) => void;
   textDecoration: string;
-//   setTextDecoration?: (textDecoration: string) => void;
   backgroundColor: string;
-//   setBackgroundColor?: (backgroundColor: string) => void;
   color: string;
-//   setColor?: (color: string) => void;
   margin: string;
-//   setMargin?: (margin: string) => void;
   width: string;
-//   setWidth?: (width: string) => void;
   height: string;
-//   setHeight?: (height: string) => void;
   padding: string;
-//   setPadding?: (padding: string) => void;
   fontSize: string;
-//   setFontSize?: (fontSize: string) => void;
   className: string;
-//   setClassName?: (className: string) => void;
 }
 
 export interface IProps {
@@ -35,29 +22,16 @@ export interface IProps {
 }
 
 export interface ISideBarProps extends IProps {
-//   id: number;
-//   element: string;  
-//   inputText: string;
   setInputText: (inputText: string) => void;
-//   texAlign: string;
   setTextAlign: (textAlign: string) => void;
-//   textDecoration: string;
   setTextDecoration: (textDecoration: string) => void;
-//   backgroundColor: string;
   setBackgroundColor: (backgroundColor: string) => void;
-//   color: string;
   setColor: (color: string) => void;
-//   margin: string;
   setMargin: (margin: string) => void;
-//   width: string;
   setWidth: (width: string) => void;
-//   height: string;
   setHeight: (height: string) => void;
-//   padding: string;
   setPadding: (padding: string) => void;
-//   fontSize: string;
   setFontSize: (fontSize: string) => void;
-//   className: string;
   setClassName: (className: string) => void;
 }
 
@@ -89,16 +63,9 @@ export interface ICustomizationProps extends IProps {
 export interface IPreviewProps { 
   elementsArr: IHtmlElement[];
   setElementsArr: (elementsArr: IHtmlElement[]) => void;
-//   previewPage?: string;
-//   setPreviewPage?: (previewPage: string) => void;
 }
 
 export interface IPreviewMainContainerProps {
   previewPage: string;
   setPreviewPage: (previewPage: string) => void;
 }
-
-// const [user, setUser] = React.useState<any>('');
-// const [projectId, setProjectId] = React.useState<any>('');
-// const [projectList, setProjectList] = React.useState([]);
-// const [loadProj, setLoadProj] = React.useState('');
