--- utils/apiKey.ts
@@ -1,5 +1,0 @@
-export const key = await crypto.subtle.generateKey(
-    { name: "HMAC", hash: "SHA-512" },
-    true,
-    ["sign", "verify"],
-  );
